pyproject.toml
[project]
name = "blog-data-analyst"
version = "0.1.0"
description = "AI-powered blog content generation and analysis agent"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Nithin Kanil" }]
keywords = ["blog", "ai", "content-generation", "seo", "langchain"]

dependencies = [
    # LLM Framework
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",  # For dual LLM quality checker
    
    # Web Research
    "browser-use>=0.1.0",
    
    # Content Extraction
    "crawl4ai>=0.3.0",
    
    # YAML Parsing
    "pyyaml>=6.0",
    
    # SEO Analysis
    "textstat>=0.7.0",
    "spacy>=3.7.0",
    
    # HTML Conversion
    "markdown>=3.5",
    "beautifulsoup4>=4.12.0",
    "pygments>=2.17.0",  # Syntax highlighting
    
    # Image Generation (Banana/Nano API)
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    
    # Data Models
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Async Support
    "asyncio>=3.4.3",
    
    # CLI
    "click>=8.1.0",
    "rich>=13.7.0",  # Pretty terminal output
    
    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
    "types-markdown>=3.5.0",
    "types-beautifulsoup4>=4.12.0",
]

[project.scripts]
blog-analyst = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
